新增实时代收商户校验，业务校验，保存实时代订单和实时代扣流水方法

+++ withhold/trade/src/main/java/com/unionpay/withhold/trade/order/pojo/MerchDetaDO.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 import java.util.Date;
 
 public class MerchDetaDO {
-    private Long merchId;
+    private Integer merchId;
 
     private Long memId;
 
@@ -66,11 +66,11 @@ public class MerchDetaDO {
 
     private String remarks;
 
-    public Long getMerchId() {
+    public Integer getMerchId() {
         return merchId;
     }
 
-    public void setMerchId(Long merchId) {
+    public void setMerchId(Integer merchId) {
         this.merchId = merchId;
     }
